feat: 统一度量衡（？）

+++ cart/src/main/java/org/liquidice/exmall/cart/dto/req/CartReqDTO.java
@@ -32,5 +32,5 @@ public class CartReqDTO {
     /**
      * 商品计算价格
      */
-    private Long productPrice;
+    private Double productPrice;
 }
