+++ order/src/main/java/org/liquidice/exmall/order/dto/req/OrderReqDTO.java
@@ -42,5 +42,5 @@ public class OrderReqDTO {
     /**
      * 订单子单价
      */
-    private Long totalPrice = 0L;
+    private Double totalPrice = 0D;
 }
