+++ pay/src/main/java/org/liquidice/exmall/pay/dto/req/PayReqDTO.java
@@ -47,5 +47,5 @@ public class PayReqDTO {
     /**
      * 支付金额
      */
-    private Long payAmount;
+    private Double payAmount;
 }
