+++ product/src/main/java/org/liquidice/exmall/product/dao/entity/ProductDO.java
@@ -39,5 +39,5 @@ public class ProductDO extends BaseDO {
     /**
      * 商品基础金额
      */
-    private Long productBaseAmount;
+    private Double productBaseAmount;
 }
