+++ product/src/main/java/org/liquidice/exmall/product/dto/req/ProductReqDTO.java
@@ -31,5 +31,5 @@ public class ProductReqDTO {
     /**
      * 商品基础金额
      */
-    private Long productBaseAmount;
+    private Double productBaseAmount;
 }
